requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Document processing
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2

# AI/ML libraries - LangChain Stack
openai>=1.10.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langsmith>=0.0.77
tiktoken>=0.5.0

# Cloud Vector Store - Pinecone - Production Ready (REST API, no gRPC)
pinecone>=3.0.0

# Cloud Vector Store - Zilliz (Milvus) - Production Ready
# pymilvus==2.3.4  # Commented out due to grpcio build issues

# Cloud Storage - Azure Blob - Production Ready  
azure-storage-blob==12.19.0

# Database - MongoDB Atlas - Production Ready
pymongo==4.6.0

# Data analysis (for analytics agent)
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
pydantic-settings>=2.1.0
httpx>=0.25.0
aiofiles>=23.2.0

# UI Framework
streamlit==1.28.1
streamlit-chat==0.1.1

# Development (optional - remove for production)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Build tools
setuptools>=65.0.0
wheel>=0.38.0
